Update expansion for HuskHomes v3.x, use gradle build system
