--- .editorconfig
@@ -1,14 +1,0 @@
-root = true
-
-[*]
-end_of_line = lf
-insert_final_newline = true
-charset = utf-8
-
-[*.{java,kt}]
-indent_style = space
-indent_size = 4
-
-[*.{yml,json}]
-indent_style = space
-indent_size = 2
