--- .gitignore
@@ -1,38 +1,0 @@
-# .gitignore designed more specifically for PlaceholderAPI expansions.
-
-# IntelliJ IDEA
-.idea/
-out/
-*.iml
-*.ipr
-*.iws
-atlassian-ide-plugin.xml
-com_crashlytics_export_strings.xml
-crashlytics.properties
-crashlytics-build.properties
-fabric.properties
-modules.xml
-
-# NetBeans
-nbproject/
-build/
-nbbuild/
-dist/
-nbdist/
-.nb-gradle/
-
-# Eclipse
-.settings/
-bin/
-tmp/
-*.tmp
-*.bak
-*.swp
-*~.nib
-.classpath
-.loadpath
-.metadata
-.project
-.recommenders
-local.properties
-/target/
